+++ Source/PlanningExtended 1.4/Source/Designators/CopyPaste/PastePlanDesignator.cs
@@ -70,7 +70,7 @@ namespace PlanningExtended.Designators
             //GenDraw.DrawNoBuildEdgeLines();
             //GenDraw.DrawFieldEdges(_selectedPlanLayout.Cells.Select(c => c.Position.ToIntVec3).ToList());
 
-            PlanManager.SetArePlansVisible(true);
+            PlanManager.SetIsPlanVisible(true);
 
             PlanLayoutUtilities.Draw(_selectedPlanLayout, UI.MouseCell(), Map);
         }